Seciton-6: Adapting components in robofriends-pwa-master to allow for more extensive testing.
