+++ projects/Section-6/robofriends-pwa-master/src/components/CounterButton.js
@@ -14,9 +14,9 @@ class CounterButton extends Component {
   }
 
   render() {
-    console.log('counter');
     return (
       <button
+        id="counter"
         color={this.props.color}
         onClick={() => this.setState(state => ({count: state.count + 1}))}>
         Count: {this.state.count}
