+++ projects/Section-6/robofriends-pwa-master/src/components/MainPage.css
@@ -1,0 +1,13 @@
+@font-face {
+    font-display: swap;
+    font-family: 'SEGA LOGO FONT';
+    font-style: normal;
+    font-weight: normal;
+    src: local('SEGA LOGO FONT'), url('../containers/SEGA.woff') format('woff');
+  }
+  
+  h1 {
+    font-family: 'SEGA LOGO FONT';
+    font-weight: 200;
+    color: #0ccac4;
+  }
