+++ projects/Section-6/robofriends-pwa-master/src/components/MainPage.test.js
@@ -1,0 +1,13 @@
+import { shallow } from 'enzyme';
+import React from 'react';
+import App from '../containers/App';
+
+test('Expect to render mainpage component',() => {
+    const moctStore = {
+        robots: [],
+        searchField: ''
+    }
+    expect(shallow(<App store={moctStore}/>)).toMatchSnapshot();
+})
+
+// redux-mock-store
