+++ projects/Section-6/robofriends-pwa-master/src/components/Scroll.js
@@ -3,7 +3,7 @@ import React from 'react';
 const Scroll = (props) => {
   console.log('Scroll');
   return (
-    <div style={{ overflow: 'scroll', border: '5px solid black', height: '800px'}}>
+    <div style={{ overflow: 'scroll', border: '5px solid black', height: '940px'}}>
       {props.children}
     </div>
   );
