--- projects/Section-6/robofriends-pwa-master/src/containers/App.test.js
@@ -1,7 +1,0 @@
-import { shallow } from 'enzyme';
-import React from 'react';
-import App from './App';
-
-test('Expect to render card component',() => {
-    
-})
